requirement.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
celery==3.1.26.post2
celery-with-redis==3.0
certifi==2019.11.28
chardet==3.0.4
ddt==1.2.1
Django==2.0
django-bootstrap==0.2.4
django-bootstrap4==0.0.8
django-celery==3.3.1
django-pure-pagination==0.3.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.12
idna==2.8
importlib-metadata==1.3.0
kombu==3.0.37
more-itertools==8.0.2
Pillow==6.2.1
PyMySQL==0.9.3
pytz==2019.1
redis==3.3.11
requests==2.22.0
six==1.12.0
urllib3==1.25.7
vine==1.3.0
xmlrunner==1.7.7
zipp==0.6.0
